fixed a couple of timing bugs and tidied up some init code. still can't get anything to display though

+++ include/timers.h
@@ -14,6 +14,6 @@ void initTimers(void);
 
 
 /* Delay for a period of time in microseconds */
-void busyWait(uint16 delay);
+void busyWait(uint32 delay);
 
 #endif
